+++ packages/nimble-components/src/table/template.ts
@@ -27,7 +27,7 @@ const isTableColumn = (): ElementsFilter => {
 export const template = html<Table>`
     <template role="table">
         <div class="table-container">
-            <div role="rowgroup">
+            <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row">
                     ${repeat(x => x.columns, html<TableColumn>`
                         <${DesignSystem.tagFor(TableHeader)} class="header">