+++ packages/nimble-components/src/wafer-map/styles.ts
@@ -2,11 +2,17 @@ import { css } from '@microsoft/fast-element';
 import { borderColor, borderWidth } from '../theme-provider/design-tokens';
 
 export const styles = css`
-    .wafer-map-container {
+    :host {
+        display: inline-block;
         width: 500px;
         height: 500px;
+    }
+
+    .wafer-map-container {
+        width: 100%;
+        padding-bottom: 100%;
         position: relative;
-        display: flex;
+        display: inline-block;
         justify-content: center;
         align-items: center;
     }
@@ -14,6 +20,7 @@ export const styles = css`
     .svg-root {
         width: 100%;
         height: 100%;
+        position: absolute;
     }
 
     .svg-root.top {
@@ -35,13 +42,14 @@ export const styles = css`
     .zoom-container {
         width: 100%;
         height: 100%;
+        position: absolute;
     }
 
     .circle-base {
         width: 100%;
         height: 100%;
         position: absolute;
-        fill: white;
+        fill: transparent;
     }
 
     .circle-drawing-path {
@@ -53,5 +61,15 @@ export const styles = css`
 
     .wafer-map-area {
         position: absolute;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        height: 100%;
+    }
+
+    .wafer-map-canvas {
+        display: inline-block;
+        width: 100%;
+        height: 100%;
     }
 `;